refactors to preoare to support date interpretations for unknowns
