+++ squidApp/src/main/java/org/cirdles/squid/gui/dateInterpretations/plots/PlotDisplayInterface.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cirdles.squid.gui.plots;
+package org.cirdles.squid.gui.dateInterpretations.plots;
 
 import java.util.List;
 import java.util.Map;
