+++ squidApp/src/main/java/org/cirdles/squid/gui/dateInterpretations/plots/squid/WeightedMeanRefreshInterface.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cirdles.squid.gui.plots.squid;
+package org.cirdles.squid.gui.dateInterpretations.plots.squid;
 
 /**
  *
